Introduce the `rules prepare` command
This commit introduces a new command for the cortextool: `rules
prepare`. It allows users to "lint" and modify their rules files to
include a specific cluster in their aggregations.

Cortex users writing from multiple systems to a single tenant is not
uncommon when you want to have a global view. In this setup, you'd
typically use a label to differentiate identical series coming from
different systems.

Cortex users transitioning from Prometheus rule evaluation to the Cortex
ruler based under this setup could do with having a tool that helps them
aggregate by an specific label.

This is exactly what this command is for.

Signed-off-by: gotjosh <josue@grafana.com>